Backend/requirements.txt
fastapi==0.110.0
uvicorn
sqlmodel==0.0.22
pydantic~=2.11.4
SQLAlchemy==2.0.38
dotenv
pydantic-settings
asyncpg
bcrypt==4.0.1
passlib
pyjwt
alembic
aioredis
redis
itsdangerous
fastapi_mail
pyotp
qrcode
Pillow